+++ src/mol-plugin/state.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -25,6 +25,8 @@ import { PluginBehavior } from './behavior';
 import { PluginCommands } from './commands';
 import { PluginConfig } from './config';
 import { PluginContext } from './context';
+import { AnimateStateSnapshotTransition } from '../mol-plugin-state/animation/built-in/state-snapshots';
+import { Scheduler } from '../mol-task';
 
 export { PluginState };
 
@@ -118,6 +120,32 @@ class PluginState extends PluginComponent {
         }
         if (snapshot.startAnimation) {
             this.animation.start();
+            return;
+        }
+
+        if (snapshot.transition?.autoplay) {
+            await Scheduler.immediatePromise();
+            this.animation.play(AnimateStateSnapshotTransition, {});
+        }
+    }
+
+    async setAnimationSnapshot(snapshot: PluginState.Snapshot, frameIndex: number) {
+        await this.animation.stopStateTransitionAnimation();
+
+        const { transition } = snapshot;
+        if (!transition) return;
+        const finalIndex = Math.min(frameIndex, transition.frames.length - 1);
+        const frame = transition.frames[finalIndex] ?? snapshot.data;
+        if (frame.data) await this.plugin.runTask(this.data.setSnapshot(frame.data));
+        if (frame.canvas3d?.props) {
+            const settings = PD.normalizeParams(Canvas3DParams, frame.canvas3d.props, 'children');
+            this.plugin.canvas3d?.setProps(settings);
+        }
+        if (frame.camera?.current) {
+            PluginCommands.Camera.Reset(this.plugin, {
+                snapshot: frame.camera.current,
+                durationMs: frame.camera.transitionStyle === 'animate' ? frame.camera.transitionDurationInMs : undefined,
+            });
         }
     }
 
@@ -211,7 +239,48 @@ namespace PluginState {
         structureComponentManager?: {
             options?: StructureComponentManager.Options
         },
-        durationInMs?: number
+        durationInMs?: number,
+        transition?: StateTransition,
+    }
+
+    export interface StateTransition {
+        autoplay?: boolean,
+        loop?: boolean,
+        frames: {
+            durationInMs: number,
+            data: State.Snapshot,
+            camera?: Snapshot['camera'],
+            canvas3d?: { props?: Canvas3DProps },
+        }[],
+    }
+
+    export function getStateTransitionDuration(snapshot: Snapshot): number | undefined {
+        const { transition } = snapshot;
+        if (!transition) return undefined;
+        let totalDuration = 0;
+        for (let i = 0; i < transition.frames.length; i++) {
+            const frame = transition.frames[i];
+            totalDuration += frame.durationInMs;
+        }
+        return totalDuration;
+    }
+
+    export function getStateTransitionFrameIndex(snapshot: Snapshot, timestamp: number): number | undefined {
+        const { transition } = snapshot;
+        if (!transition) return undefined;
+
+        let t = timestamp;
+        if (transition.loop) {
+            t %= getStateTransitionDuration(snapshot) ?? 1;
+        }
+
+        let currentDuration = 0;
+        for (let i = 0; i < transition.frames.length; i++) {
+            if (currentDuration >= t) return i;
+            const frame = transition.frames[i];
+            currentDuration += frame.durationInMs;
+        }
+        return transition.frames.length - 1;
     }
 
     export type SnapshotType = 'json' | 'molj' | 'zip' | 'molx'