+++ src/mol-repr/volume/dot.ts
@@ -2,6 +2,7 @@
  * Copyright (c) 2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Gianluca Tomasello <giagitom@gmail.com>
  */
 
 import { ParamDefinition as PD } from '../../mol-util/param-definition';
@@ -26,9 +27,11 @@ import { addSphere } from '../../mol-geo/geometry/mesh/builder/sphere';
 import { sphereVertexCount } from '../../mol-geo/primitive/sphere';
 import { Points } from '../../mol-geo/geometry/points/points';
 import { PointsBuilder } from '../../mol-geo/geometry/points/points-builder';
+import { Mat4 } from '../../mol-math/linear-algebra';
 
 export const VolumeDotParams = {
     isoValue: Volume.IsoValueParam,
+    perturbPositions: PD.Boolean(false)
 };
 export type VolumeDotParams = typeof VolumeDotParams
 export type VolumeDotProps = PD.Values<VolumeDotParams>
@@ -58,9 +61,10 @@ export function VolumeSphereImpostorVisual(materialId: number): VolumeVisual<Vol
         createLocationIterator: createVolumeCellLocationIterator,
         getLoci: getDotLoci,
         eachLocation: eachDot,
-        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumeSphereParams>, currentProps: PD.Values<VolumeSphereParams>) => {
+        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumeSphereParams>, currentProps: PD.Values<VolumeSphereParams>, newTheme: Theme, currentTheme: Theme) => {
             state.createGeometry = (
-                !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats)
+                !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats) ||
+                newProps.perturbPositions !== currentProps.perturbPositions
             );
         },
         geometryUtils: Spheres.Utils,
@@ -77,9 +81,10 @@ export function VolumeSphereMeshVisual(materialId: number): VolumeVisual<VolumeS
         createLocationIterator: createVolumeCellLocationIterator,
         getLoci: getDotLoci,
         eachLocation: eachDot,
-        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumeSphereParams>, currentProps: PD.Values<VolumeSphereParams>) => {
+        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumeSphereParams>, currentProps: PD.Values<VolumeSphereParams>, newTheme: Theme, currentTheme: Theme) => {
             state.createGeometry = (
                 !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats) ||
+                newProps.perturbPositions !== currentProps.perturbPositions ||
                 newProps.sizeFactor !== currentProps.sizeFactor ||
                 newProps.detail !== currentProps.detail
             );
@@ -91,6 +96,26 @@ export function VolumeSphereMeshVisual(materialId: number): VolumeVisual<VolumeS
     }, materialId);
 }
 
+type Basis = { x: Vec3, y: Vec3, z: Vec3, maxScale: number }
+function getBasis(m: Mat4): Basis {
+    return {
+        ...Mat4.extractBasis(m),
+        maxScale: Mat4.getMaxScaleOnAxis(m)
+    };
+}
+
+const offset = Vec3();
+function getRandomOffsetFromBasis({ x, y, z, maxScale }: Basis): Vec3 {
+    const rx = (Math.random() - 0.5) * maxScale;
+    const ry = (Math.random() - 0.5) * maxScale;
+    const rz = (Math.random() - 0.5) * maxScale;
+
+    Vec3.scale(offset, x, rx);
+    Vec3.scaleAndAdd(offset, offset, y, ry);
+    Vec3.scaleAndAdd(offset, offset, z, rz);
+
+    return offset;
+}
 
 export function createVolumeSphereImpostor(ctx: VisualContext, volume: Volume, key: number, theme: Theme, props: VolumeSphereProps, spheres?: Spheres): Spheres {
     const { cells: { space, data }, stats } = volume.grid;
@@ -103,14 +128,28 @@ export function createVolumeSphereImpostor(ctx: VisualContext, volume: Volume, k
     const count = Math.ceil((xn * yn * zn) / 10);
     const builder = SpheresBuilder.create(count, Math.ceil(count / 2), spheres);
 
+    const invert = isoVal < 0;
+
+    // Precompute basis vectors and largest cell axis length
+    const basis = props.perturbPositions ? getBasis(gridToCartn) : undefined;
+
     for (let z = 0; z < zn; ++z) {
         for (let y = 0; y < yn; ++y) {
             for (let x = 0; x < xn; ++x) {
-                if (space.get(data, x, y, z) < isoVal) continue;
+                const value = space.get(data, x, y, z);
+                if (!invert && value < isoVal || invert && value > isoVal) continue;
 
-                Vec3.set(p, x, y, z);
-                Vec3.transformMat4(p, p, gridToCartn);
-                builder.add(p[0], p[1], p[2], space.dataOffset(x, y, z));
+                const cellIdx = space.dataOffset(x, y, z);
+                if (basis) {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                    const offset = getRandomOffsetFromBasis(basis);
+                    Vec3.add(p, p, offset);
+                } else {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                }
+                builder.add(p[0], p[1], p[2], cellIdx);
             }
         }
     }
@@ -130,24 +169,37 @@ export function createVolumeSphereMesh(ctx: VisualContext, volume: Volume, key: 
     const p = Vec3();
     const [xn, yn, zn] = space.dimensions;
 
-    const count = (xn * yn * zn) / 10;
+    const count = Math.ceil((xn * yn * zn) / 10);
     const vertexCount = count * sphereVertexCount(detail);
-    const builderState = MeshBuilder.createState(vertexCount, vertexCount / 2, mesh);
+    const builderState = MeshBuilder.createState(vertexCount, Math.ceil(vertexCount / 2), mesh);
 
     const l = Volume.Cell.Location(volume);
     const themeSize = theme.size.size;
+    const invert = isoVal < 0;
+
+    // Precompute basis vectors and largest cell axis length
+    const basis = props.perturbPositions ? getBasis(gridToCartn) : undefined;
 
     for (let z = 0; z < zn; ++z) {
         for (let y = 0; y < yn; ++y) {
             for (let x = 0; x < xn; ++x) {
-                if (space.get(data, x, y, z) < isoVal) continue;
+                const value = space.get(data, x, y, z);
+                if (!invert && value < isoVal || invert && value > isoVal) continue;
 
-                Vec3.set(p, x, y, z);
-                Vec3.transformMat4(p, p, gridToCartn);
-                builderState.currentGroup = space.dataOffset(x, y, z);
-                l.cell = builderState.currentGroup as Volume.CellIndex;
-                const size = themeSize(l);
-                addSphere(builderState, p, size * sizeFactor, detail);
+                const cellIdx = space.dataOffset(x, y, z);
+                l.cell = cellIdx as Volume.CellIndex;
+                const size = themeSize(l) * sizeFactor;
+                if (basis) {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                    const offset = getRandomOffsetFromBasis(basis);
+                    Vec3.add(p, p, offset);
+                } else {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                }
+                builderState.currentGroup = cellIdx;
+                addSphere(builderState, p, size, detail);
             }
         }
     }
@@ -173,9 +225,10 @@ export function VolumePointVisual(materialId: number): VolumeVisual<VolumePointP
         createLocationIterator: createVolumeCellLocationIterator,
         getLoci: getDotLoci,
         eachLocation: eachDot,
-        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumePointParams>, currentProps: PD.Values<VolumePointParams>) => {
+        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<VolumePointParams>, currentProps: PD.Values<VolumePointParams>, newTheme: Theme, currentTheme: Theme) => {
             state.createGeometry = (
-                !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats)
+                !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats) ||
+                newProps.perturbPositions !== currentProps.perturbPositions
             );
         },
         geometryUtils: Points.Utils,
@@ -193,14 +246,28 @@ export function createVolumePoint(ctx: VisualContext, volume: Volume, key: numbe
     const count = Math.ceil((xn * yn * zn) / 10);
     const builder = PointsBuilder.create(count, Math.ceil(count / 2), points);
 
+    const invert = isoVal < 0;
+
+    // Precompute basis vectors and largest cell axis length
+    const basis = props.perturbPositions ? getBasis(gridToCartn) : undefined;
+
     for (let z = 0; z < zn; ++z) {
         for (let y = 0; y < yn; ++y) {
             for (let x = 0; x < xn; ++x) {
-                if (space.get(data, x, y, z) < isoVal) continue;
+                const value = space.get(data, x, y, z);
+                if (!invert && value < isoVal || invert && value > isoVal) continue;
 
-                Vec3.set(p, x, y, z);
-                Vec3.transformMat4(p, p, gridToCartn);
-                builder.add(p[0], p[1], p[2], space.dataOffset(x, y, z));
+                const cellIdx = space.dataOffset(x, y, z);
+                if (basis) {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                    const offset = getRandomOffsetFromBasis(basis);
+                    Vec3.add(p, p, offset);
+                } else {
+                    Vec3.set(p, x, y, z);
+                    Vec3.transformMat4(p, p, gridToCartn);
+                }
+                builder.add(p[0], p[1], p[2], cellIdx);
             }
         }
     }